add flag lint rules
