+++ internal/cmd/kafka/command_acl_delete.go
@@ -23,11 +23,22 @@ func (c *aclCommand) newDeleteCommand() *cobra.Command {
 		RunE:  c.delete,
 	}
 
-	cmd.Flags().AddFlagSet(aclConfigFlags())
+	cmd.Flags().StringSlice("operations", []string{""}, fmt.Sprintf("A comma-separated list of ACL operations: (%s).", listEnum(ccstructs.ACLOperations_ACLOperation_name, []string{"ANY", "UNKNOWN"})))
+	cmd.Flags().String("principal", "", `Principal for this operation, prefixed with "User:".`)
+	cmd.Flags().String("service-account", "", "The service account ID.")
+	cmd.Flags().Bool("allow", false, "Access to the resource is allowed.")
+	cmd.Flags().Bool("deny", false, "Access to the resource is denied.")
+	cmd.Flags().Bool("cluster-scope", false, "Modify ACLs for the cluster.")
+	cmd.Flags().String("topic", "", "Modify ACLs for the specified topic resource.")
+	cmd.Flags().String("consumer-group", "", "Modify ACLs for the specified consumer group resource.")
+	cmd.Flags().String("transactional-id", "", "Modify ACLs for the specified TransactionalID resource.")
+	cmd.Flags().Bool("prefix", false, "When this flag is set, the specified resource name is interpreted as a prefix.")
 	pcmd.AddForceFlag(cmd)
 	pcmd.AddClusterFlag(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddContextFlag(cmd, c.CLICommand)
 	pcmd.AddEnvironmentFlag(cmd, c.AuthenticatedCLICommand)
+
+	_ = cmd.MarkFlagRequired("operations")
 
 	return cmd
 }
