+++ internal/cmd/schema-registry/command_subject_update_onprem.go
@@ -21,11 +21,11 @@ func (c *subjectCommand) newUpdateCommandOnPrem() *cobra.Command {
 		Example: examples.BuildExampleString(
 			examples.Example{
 				Text: `Update subject level compatibility of subject "payments"`,
-				Code: fmt.Sprintf("%s schema-registry subject update payments --compatibility=BACKWARD %s", version.CLIName, OnPremAuthenticationMsg),
+				Code: fmt.Sprintf("%s schema-registry subject update payments --compatibility BACKWARD %s", version.CLIName, OnPremAuthenticationMsg),
 			},
 			examples.Example{
 				Text: `Update subject level mode of subject "payments"`,
-				Code: fmt.Sprintf("%s schema-registry subject update payments --mode=READWRITE %s", version.CLIName, OnPremAuthenticationMsg),
+				Code: fmt.Sprintf("%s schema-registry subject update payments --mode READWRITE %s", version.CLIName, OnPremAuthenticationMsg),
 			},
 		),
 	}
