--- test/fixtures/output/cluster/list-type-unknown.golden
@@ -1,12 +1,0 @@
-Error: type should be one of connect-cluster, kafka-cluster, ksql-cluster, schema-registry-cluster
-Usage:
-  confluent cluster list [flags]
-
-Flags:
-      --type string     Filter list to this cluster type (connect-cluster, kafka-cluster, ksql-cluster, schema-registry-cluster).
-  -o, --output string   Specify the output format as "human", "json", or "yaml". (default "human")
-  -h, --help            Show help for this command.
-
-Global Flags:
-  -v, --verbose count   Increase verbosity (-v for warn, -vv for info, -vvv for debug, -vvvv for trace).
-
