+++ test/fixtures/output/schema-registry/subject/update-help.golden
@@ -6,11 +6,11 @@ Usage:
 Examples:
 Update subject level compatibility of subject "payments".
 
-  $ confluent schema-registry subject update payments --compatibility=BACKWARD
+  $ confluent schema-registry subject update payments --compatibility BACKWARD
 
 Update subject level mode of subject "payments".
 
-  $ confluent schema-registry subject update payments --mode=READWRITE
+  $ confluent schema-registry subject update payments --mode READWRITE
 
 Flags:
       --compatibility string   Can be BACKWARD, BACKWARD_TRANSITIVE, FORWARD, FORWARD_TRANSITIVE, FULL, FULL_TRANSITIVE, or NONE.
